moved addons for web application scanners to new sub-folder of examples/complex.
